update IUC, ImportToDb method in Model

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -18,7 +18,7 @@ class CreateUsersTable extends Migration
             $table->string('Name');
             $table->string('Age');
             $table->string('Email');
-            $table->decimal('Phone');
+            $table->bigInteger('Phone');
             $table->string('Gender');
         });
     }
